api/requirements.txt
tensorflow==2.12.0
# tensorflow-serving-api removed (it conflicts badly)
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy
